Initial commit
The skeleton for a language syntax extension was created with the "yo
code" generator, pointing it to the URL with the TextMate syntax file
for Acton. You can build the .vsix file with "vsce package" and then
install the extension in your local VS Code.

Syntax highlighting works, but the Acton language is not automatically
selected for .act files, even though it is listed in package.json?!