Enable build & run
We have commands; Acton: Build Active File, Acton: Run Active File to just
build or build and run the active file. We correctly move to the proper working
directory for project files or standalone files.

The Run (and debug) button in VS Code also works to run the program, using the
Build Active file command and then running the program. We rely on lldb-dap for
this but do not wire up any actual debugging. stdout from the process is
printed to the Debug Console.

lldb-dap is added as a dependency and should be automatically installed.